Fixed everything but the activity tests. App now running.
